+++ roboclaw/__init__.py
@@ -1,0 +1,3 @@
+"""module management for the RoboClaw package"""
+from .roboclaw_3 import Roboclaw
+__all__ = ['Roboclaw']
